changed to new view

+++ src/screens/ArtPreviewScreen.js
@@ -19,6 +19,9 @@ import CommentsModal from "../assets/components/CommentsModal";
 import { AntDesign, Entypo, FontAwesome, FontAwesome5, Fontisto } from "@expo/vector-icons";
 import LoaderImage from "../assets/components/LoaderImage";
 import { Alert } from "react-native";
+import TransparentHeaderView from "../components/TransparentHeaderView";
+import { ActionButton, ArtDetails, ArtInfoCard, ArtistArtworksCard, ArtworkImageSlider, UserActivityCard, ViewAll } from "../components";
+import { Cart } from "../components/icons";
 
 // import Toast from "react-native-simple-toast";
 // line 141 (image is a string) and 146 (catch block)
@@ -29,12 +32,12 @@ export default function ArtPreviewScreen({ route, navigation }) {
   const [isLiked, setIsLiked] = useState(false);
   const [following, setFollowing] = useState("");
   const [image, setImage] = useState("");
-  const [uuid, setUId] = useState(auth.currentUser.uid);
+  const [uid, setUId] = useState(null);
   const [currentUserLike, setCurrentUserLike] = useState(false);
   const [photoURL, setPhotoURL] = useState(null);
   const [FullName, setFullName] = useState(null);
   const [artistDescription, setArtistDescription] = useState("");
-  const [artistName, setArtistName] = useState("");
+  // const [artistName, setArtistName] = useState("");
   const [artistPhoto, setArtistPhoto] = useState("");
   const [artType, setArtType] = useState("");
   const [price, setPrice] = useState(0);
@@ -47,10 +50,54 @@ export default function ArtPreviewScreen({ route, navigation }) {
   const [defaultImg] = useState('https://via.placeholder.com/150/0000FF/808080%20?Text=Digital.comC/O%20https://placeholder.com/')
   const [artHeight, setArtHeight] = useState(400);
   const [artWidth, setArtWidth] = useState(400);
+
+  // new design details
+  const [artDetails, setArtDetails] = useState({ price: null, description: null, dimensions: null, artName: '', imgUrls: [] })
+  const [artworks, setArtworks] = useState([])
   // console.log({ ...route.params });
-  const { artistUid, imageUID } = route.params;
+  const { artistUid, imageUID, photoUrl, artistName } = route.params;
 
   // const [Data] = useState([{ photoURL, FullName, imageUID, title: 'why' }, { photoURL, FullName, imageUID, title: 'why' }, { photoURL, FullName, imageUID, title: 'why' }])
+  useEffect(() => {
+    console.log({ artistUid, imageUID, photoUrl });
+    getArtDetails();
+    getArt()
+  }, [route])
+
+  useEffect(() => {
+    auth.onAuthStateChanged(user => {
+      console.log({ user });
+      if (!user) {
+        console.log('No user found');
+        return
+      }
+      setUId(user.uid)
+    })
+  }, [])
+  useEffect(() => {
+    console.log({ uid });
+  }, [uid])
+
+  const getArt = async () => {
+    // console.log(`artist ID : ${artistUid}s`);
+    console.log('running art getter');
+    firestore
+      .collection("newArtworks")
+      .where("userid", "==", artistUid)
+      .where("isEnabled", "==", true)
+      .onSnapshot((snapShot) => {
+        if (!snapShot.empty) {
+          let allArt = snapShot.docs.map((docSnap) => ({ isArt: true, ...docSnap.data() }));
+          // console.log('all art: ', allArt);
+          // if (allArt.length > 0) {
+          //   setArt([...allArt, { isArt: false, ImageUid: null }]);
+          // }
+          setArtworks(allArt)
+        }
+
+
+      });
+  };
   useEffect(() => {
     try {
       (async () => {
@@ -92,9 +139,9 @@ export default function ArtPreviewScreen({ route, navigation }) {
   };
 
   const onLikePress = () => {
-    const uid = auth.currentUser.uid;
+    // const uid = auth.currentUser.uid;
     firestore
-      .collection("Market")
+      .collection("newArtworks")
       .doc(imageUID)
       .collection("likes")
       .doc(uid)
@@ -111,9 +158,10 @@ export default function ArtPreviewScreen({ route, navigation }) {
   };
 
   const onDislikePress = () => {
-    const uid = auth.currentUser.uid;
+    // const uid = auth.currentUser.uid;
+    if (!imageUID) return
     firestore
-      .collection("Market")
+      .collection("newArtworks")
       .doc(imageUID)
       .collection("likes")
       .doc(uid)
@@ -127,11 +175,12 @@ export default function ArtPreviewScreen({ route, navigation }) {
   const likesState = () => {
     try {
       // const uid = auth.currentUser.uid;
+      if (!imageUID) return
       firestore
-        .collection("Market")
+        .collection("newArtworks")
         .doc(imageUID)
         .collection("likes")
-        .where("uid", "==", uuid)   //change this to fetch one document, use uid to add likes
+        .where("uid", "==", uid)   //change this to fetch one document, use uid to add likes
         .onSnapshot((snapShot) => {
           if (snapShot.exists) {
             const imag = snapShot.docs
@@ -156,8 +205,9 @@ export default function ArtPreviewScreen({ route, navigation }) {
   const getNumberOfLikes = () => {
     try {
       // add catch block to handle errors
+      if (!imageUID) return
       firestore
-        .collection("MarKet")
+        .collection("newArtworks")
         .doc(imageUID)
         .collection("likes")
         .where("artistUid", "==", artistUid)
@@ -177,11 +227,15 @@ export default function ArtPreviewScreen({ route, navigation }) {
   };
 
   const getArtDetails = () => {
+    if (!imageUID) return
     firestore
-      .collection("Market")
+      .collection("newArtworks")
       .doc(imageUID)
       .onSnapshot((snapShot) => {
         if (snapShot.exists) {
+          console.log({ data: snapShot.data() });
+          setArtDetails({ ...snapShot.data(), images: snapShot.data().imgUrls.map(item => item.imgUrl)})
+          return
           const artTypes = snapShot.data().artType;
           setArtType(artTypes);
           const prices = snapShot.data().price;
@@ -197,11 +251,14 @@ export default function ArtPreviewScreen({ route, navigation }) {
         }
       });
   };
-
+  useEffect(() => { 
+    console.log({ artDetailsInPreview: artDetails });
+  }, [artDetails])
   const addToCart = async () => {
     // console.log('adding to cart');
     // return
-    const uid = auth.currentUser.uid;
+    // const uid = auth.currentUser.uid;
+    if (!imageUID) return
     await firestore
       .collection("cartItem")
       .doc(uid)
@@ -211,7 +268,7 @@ export default function ArtPreviewScreen({ route, navigation }) {
         artUrl: artUrl,
         artType: artType,
         price: price,
-        uuid: uuid,
+        uid: uid,
         artistUid: artistUid,
         imageUid: imageUID,
       })
@@ -228,8 +285,9 @@ export default function ArtPreviewScreen({ route, navigation }) {
       });
   };
   const removeFromCart = async () => {
-    const uid = auth.currentUser.uid;
+    // const uid = auth.currentUser.uid;
     try {
+      if (!imageUID) return
       await firestore.collection('cartItem').doc(uid).collection('items').doc(imageUID).delete().then((res) => {
         // console.log(res);
         setItemOnCart(false)
@@ -256,9 +314,9 @@ export default function ArtPreviewScreen({ route, navigation }) {
   };
 
   const onFollowing = async () => {
-    const uuid = auth.currentUser.uid;
+    // const uuid = auth.currentUser.uid;
     const update = {
-      uuid: uuid,
+      uid: uid,
       artistUid: artistUid,
       photo: photoURL,
       artistPhoto: artistPhoto,
@@ -266,10 +324,10 @@ export default function ArtPreviewScreen({ route, navigation }) {
       artistName: artistName,
     }
     try {
-      const res = await firestore.collection('following').doc(artistUid).collection('userFollowing').doc(uuid).set(update)
-        // console.log('res: ');
-      if(res) setFollowing(true)
-      
+      const res = await firestore.collection('following').doc(artistUid).collection('userFollowing').doc(uid).set(update)
+      // console.log('res: ');
+      if (res) setFollowing(true)
+
       // console.log(status);
     } catch (error) {
       console.log(error);
@@ -279,13 +337,13 @@ export default function ArtPreviewScreen({ route, navigation }) {
   const onUnFollowing = async () => {
     // setFollowing(false);
     // return
-    const uuid = auth.currentUser.uid;
+    // const uuid = auth.currentUser.uid;
 
     return firestore
       .collection("following")
       .doc(artistUid)
       .collection("userFollowing")
-      .doc(uuid)
+      .doc(uid)
       .delete()
       .then(() => {
         try {
@@ -310,7 +368,7 @@ export default function ArtPreviewScreen({ route, navigation }) {
     } catch (error) {
       console.log(error);
     }
-    
+
   }
   const followState = () => {
     try {
@@ -321,7 +379,7 @@ export default function ArtPreviewScreen({ route, navigation }) {
           if (snapShot1.exists) {
             snapShot1.ref
               .collection("userFollowing")
-              .where("uuid", "==", uid)
+              .where("uid", "==", uid)
               .get().then(snapShot => {
                 snapShot.docs.map((document) => {
                   if (document.exists) {
@@ -339,7 +397,7 @@ export default function ArtPreviewScreen({ route, navigation }) {
         { text: 'Ok', onPress: () => console.log('okay pressed') }
       ])
     }
-    const uid = auth.currentUser.uid;
+    // const uid = auth.currentUser.uid;
   };
 
   const disableObjects = async () => {
@@ -347,56 +405,66 @@ export default function ArtPreviewScreen({ route, navigation }) {
       setDisplayContent(!displayContent);
     }, 250);
   };
+  const updateState = async (item, imageUID) => {
+    console.log({ item, imageUID });
+    navigation.navigate('ArtPreview', { artistUid, imageUID: item.ImageUid, photoUrl, artistName })
+  }
 
   useEffect(() => {
-    const uid = auth.currentUser.uid;
-    let isMounted = true;
-    try {
-      if (isMounted) {
-        firestore
-          .collection("users")
-          .doc(uid)
-          .onSnapshot((snapShot) => {
-            if (snapShot.exists) {
-              const users = snapShot.data().photoURL;
-              const uName = snapShot.data().fullName;
-              console.log({ users, uName });
-              setPhotoURL(users);
-              setFullName(uName);
-            }
-          });
+    console.log({ route });
+  }, [route])
 
-        getNumberOfLikes();
+  // useEffect(() => {
+  //   // const uid = auth.currentUser.uid;
+  //   let isMounted = true;
+  //   try {
+  //     if (isMounted) {
+  //       if(!imageUID) return
+  //       firestore
+  //         .collection("users")
+  //         .doc(uid)
+  //         .onSnapshot((snapShot) => {
+  //           if (snapShot.exists) {
+  //             const users = snapShot.data().photoURL;
+  //             const uName = snapShot.data().fullName;
+  //             console.log({ users, uName });
+  //             setPhotoURL(users);
+  //             setFullName(uName);
+  //           }
+  //         });
 
-        likesState();
-        followState();
-      }
-    } catch (error) {
-      Alert.alert('error', error, [
-        { text: 'Ok', onPress: () => console.log('okay pressed') }
-      ])
-    }
+  //       getNumberOfLikes();
 
-    return () => isMounted = false;
-  }, [imageUID, artistUid]);
-  useEffect(() => {
-    let isMounted = true;
-    if (isMounted) {
-      const uid = auth.currentUser.uid
-      firestore.collection('cartItem').doc(uid).collection('items').doc(imageUID).get().then((res) => {
-        // console.log(res.data());4
-        console.log(imageUID);
-        if (res.exists) {
-          // console.log(res.data());
-          setItemOnCart(true)
-        }
-      }).catch(err => console.log(err))
+  //       likesState();
+  //       followState();
+  //     }
+  //   } catch (error) {
+  //     Alert.alert('error', error, [
+  //       { text: 'Ok', onPress: () => console.log('okay pressed') }
+  //     ])
+  //   }
 
-      getArtistDetailts();
-      getArtDetails();
-    }
-    return () => isMounted = false;
-  }, [])
+  //   return () => isMounted = false;
+  // }, [imageUID, artistUid]);
+  // useEffect(() => {
+  //   let isMounted = true;
+  //   if (isMounted) {
+  //     // const uid = auth.currentUser.uid
+  //     if(!imageUID) return
+  //     firestore.collection('cartItem').doc(uid).collection('items').doc(imageUID).get().then((res) => {
+  //       // console.log(res.data());4
+  //       console.log(imageUID);
+  //       if (res.exists) {
+  //         // console.log(res.data());
+  //         setItemOnCart(true)
+  //       }
+  //     }).catch(err => console.log(err))
+
+  //     getArtistDetailts();
+  //     getArtDetails();
+  //   }
+  //   return () => isMounted = false;
+  // }, [])
 
   const RenderScrollView = ({ photoURL, FullName, imageUID, style }) => {
     // console.log(imageUID);
@@ -645,20 +713,61 @@ export default function ArtPreviewScreen({ route, navigation }) {
 
   }
   return (
-    <RenderScrollView
-      photoURL={photoURL}
-      FullName={FullName}
-      imageUID={imageUID}
-      style={{
-        borderColor: 'red',
-        borderWidth: 1,
-        alignSelf: 'center',
-        flex: 1,
-        backgroundColor: 'red',
-        height: Dimensions.get('window').height,
-        width: Dimensions.get('window').width
-      }}
-    />
+    <TransparentHeaderView padding={0}>
+      <ScrollView >
+        <ArtworkImageSlider imagesArr={artDetails.images} />
+        <View style={{ paddingHorizontal: 15, gap: 10 }}>
+          <ArtDetails price={artDetails.price} description={artDetails.description} artName={artDetails.artName} />
+          <ArtInfoCard condition={artDetails.condition} available={artDetails.isAvailable} dimensions={artDetails.dimensions} />
+          <ActionButton icon={<Cart size={24} />} text={'Add to Cart'} disabled={false} onPress={() => {addToCart()}} style={{ marginBottom: 20 }} />
+          <UserActivityCard artistName={artistName} following={true} likes={30} messages={30} artistPhoto={photoUrl} viewArtist={() => navigation.navigate('ArtistProfile', {artistUid, photoUrl, artistName})}/>
+          <View style={{ padding: 0 }}>
+            <View style={styles.moreArtHeader}>
+              <Text style={styles.text}>
+                More from {artistName}
+              </Text>
+              <ViewAll onPress={() => console.log('View More from preview screen selected')} />
+            </View>
+            {
+              artworks && artworks.length > 0 ? (
+                <View
+                  style={{ width: Dimensions.get('window').width, left: -15, paddingHorizontal: 20, backgroundColor: 'blue', overflow: 'visible' }}
+                >
+                  <FlatList
+                    horizontal
+                    style={{ width: '100%', paddingHorizontal: 0, paddingVertical: 10, backgroundColor: 'red', gap: 8, overflow: 'visible' }}
+                    // numColumns={2}
+                    scrollEnabled
+                    columnWrapperStyle={styles.columnWrapper}
+                    ItemSeparatorComponent={() => <View style={{ width: 20 }} />}
+                    data={artworks}
+                    renderItem={({ item }) => (
+                      <ArtistArtworksCard
+                        imageUID={item.ImageUid}
+                        onPress={(imageUID) => updateState(item, imageUID)}
+                        showPrice={true}
+                        artistName={artistName}
+                        artName={item.artName}
+                        artUri={item.artUrl}
+                        artistPic={photoUrl}
+                        price={item.price}
+                      />
+                    )}
+                    keyExtractor={item => item.ImageUid}
+                  />
+                </View>
+
+              ) : (
+                <View>
+                  <Text>No more artwork</Text>
+                </View>
+              )
+            }
+          </View>
+        </View>
+
+      </ScrollView>
+    </TransparentHeaderView>
   )
 
 }
@@ -678,5 +787,14 @@ const styles = StyleSheet.create({
   },
   tikTokView: {
     height: Dimensions.get('window').height
+  },
+  moreArtHeader: {
+    flexDirection: 'row',
+    justifyContent: 'space-between',
+    alignItems: 'center',
+    width: '100%',
+    // paddingHori,
+    backgroundColor: 'blue'
   }
+
 });
